requirements.txt
flask==3.1.2
supabase==2.18.1
python-dotenv==3.12.3
werkzeug==3.1.3
gunicorn==21.2.0
auth0-python==4.2.0
bcrypt==4.3.0